Remove function runInstallationScriptIfNecessary.
The function was used for bootstrapping in the assign command.
It was created to install packages in the vendor folder thus creating
the vendor folder.  A circular dependency was created, since the
autoload script looks for the vendor folder and the function was
requiring the autoload script.  Installation of packages should be
done before including the autoload script.