Change the default parameter value of the third argument of the registerFile function.
Making the default filename of the registry the default parameter value
simplifies the logic in the function by removing unnecessary null checks.
This commit fixes a bug that caused registry to not be generated if it
did not exist.  The bug was caused by not doing a file exists check
when the default filename was used.  The function now does a file exists
check whether the default filename is used or one is given.