Add sub method to Date class.
It will be best to use this method when creating new instances
minus 'n' days.  The primitives used to construct Date instances
do not contain enough internal state to derive reliable subtractions.
For example, if a 'May 1st' DayOfMonth instance were used to derive
an instance with one day subtracted, a 'May 30th' DayOfMonth instance
would be returned since the class does not contain enough internal state.

It's best to do such calculations on instances of Date instead of the
primitives anyway.

This change was made as a result of a failure in
SpecialEventTest::testAccessorsReturnExpectedValues.