Create backup and restore feature.
Use to migrate or recover data, contacts, mail_configuration, etc. to
another location (local or remote).  A password protected
backup file is created.  The path to the backup file is used
to restore.  The restore feature does not download/upload the
backup file.  The base directory of the target app is used
as the root directory for "restoring" files.