Stop using contact's id as assignment form filename.
In order to strengthen privacy policies, the sha1 of the intended
recipient's id will be used as the filename of the assignment form
that will be sent.  Better error handling has also been put in
place which takes care of occasions when the filename can not be
mapped to anyone in the list of contacts or when the email service
throws an exception.