Fix an error that occurs when creating Month objects.
The error occurs on the last day of a month containing 31 days,
which may cause an overflow when a Month instance is created.
This is because PHP's DateTime object allows for creating instances
using a month without giving the day of the month.  When the day
of the month is not given, the constructor will use the current
day of the month.  If the current day is the 31st and an attempt
is made to create a DateTime object for a month with fewer than
31 days, an instance will be created representing the month after
what was intended.  The first day of the month is used when creating
Month objects.
https://www.php.net/manual/en/datetime.createfromformat.php

Revert "Add middle name compatibility to contact and fullname instances."

This reverts commit f1a7ab5d13dd934d7399253e2c9ccb130cce5f29.